- Use the console subsystem on Windows for now. - Fixed: A really small subset of the warnings produced by MSVC.

+++ src/thingdef/thingdef_expression.h
@@ -51,11 +51,11 @@ class ExpressionNode
 			public:
 				void PerformOperation(const Value *other, const ExpressionOperator &op);
 
-				const Value &operator=(int64_t val) { i = val; d = val; isDouble = false; return *this; }
-				const Value &operator=(double val) { i = val; d = val; isDouble = true; return *this; }
+				const Value &operator=(int64_t val) { i = val; d = static_cast<double> (val); isDouble = false; return *this; }
+				const Value &operator=(double val) { i = static_cast<int64_t>(val); d = val; isDouble = true; return *this; }
 
-				int64_t GetInt() const { return isDouble ? d : i; }
-				double GetDouble() const { return isDouble ? d : i; }
+				int64_t GetInt() const { return isDouble ? static_cast<int64_t>(d) : i; }
+				double GetDouble() const { return isDouble ? d : static_cast<double>(i); }
 
 			private:
 				bool isDouble;
